requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
python-dotenv
mistralai
requests
pymysql
PyPDF2
